+++ bookstore-api/services/book.service.js
@@ -1,4 +1,5 @@
 import BookRepository from "../repositories/book.repository.js";
+import saleRepository from "../repositories/sale.repository.js";
 
 async function createBook(book) {
   await BookRepository.createBook(book);
@@ -9,18 +10,20 @@ async function updateBook(book) {
 }
 
 async function deleteBook(id) {
-  // const animal = await AnimalRepository.getAnimalByBookId(id);
-  // if (animal) {
-  //   throw new Error(
-  //     "Forbidden deletion! There are animals registered for this book!"
-  //   );
-  // } else {
-  //   return await BookRepository.deleteBook(id);
-  // }
-  return await BookRepository.deleteBook(id);
+  const sales = await saleRepository.getSaleByBookId(id);
+  if (sales.length !== 0) {
+    throw new Error(
+      "Forbidden deletion! There are sales registered for this book!"
+    );
+  } else {
+    return await BookRepository.deleteBook(id);
+  }
 }
 
-async function getBooks() {
+async function getBooks(authorId) {
+  if (authorId) {
+    return await BookRepository.getBookByAuthorId(authorId);
+  }
   return await BookRepository.getBooks();
 }
